modify the join call, not my_address
