+++ chain-signatures/node/src/protocol/contract/primitives.rs
@@ -8,7 +8,6 @@ use std::{
     hash::Hash,
     str::FromStr,
 };
-use url;
 
 type ParticipantId = u32;
 
@@ -59,7 +58,7 @@ impl From<mpc_contract::primitives::Participants> for Participants {
                                 contract_participant_info.account_id.as_ref(),
                             )
                             .unwrap(),
-                            url: Self::process_url(contract_participant_info.url),
+                            url: util::process_url(contract_participant_info.url),
                             cipher_pk: hpke::PublicKey::from_bytes(
                                 &contract_participant_info.cipher_pk,
                             ),
@@ -88,7 +87,7 @@ impl From<Candidates> for Participants {
                         ParticipantInfo {
                             id: participant_id as ParticipantId,
                             account_id,
-                            url: Self::process_url(candidate_info.url),
+                            url: util::process_url(candidate_info.url),
                             cipher_pk: candidate_info.cipher_pk,
                             sign_pk: candidate_info.sign_pk,
                         },
@@ -198,28 +197,6 @@ impl Participants {
         Participants {
             participants: intersect,
         }
-    }
-
-    /// Process the URL to remove the port if it is not an IP address.
-    /// This is needed because some mainnet partners joined with a port number included in their url
-    /// and we need to remove that port in order to connect to the node.
-    fn process_url(url_str: String) -> String {
-        let url_str = url_str.clone();
-        url::Url::parse(&url_str)
-            .map(|mut url| {
-                if !util::is_url_host_ip(&url) {
-                    if let Err(err) = url.set_port(None) {
-                        tracing::warn!(
-                            "Error setting participant's url {url} port to None: {err:?}"
-                        );
-                        return url_str.clone();
-                    }
-                    url.to_string()
-                } else {
-                    url_str.clone()
-                }
-            })
-            .unwrap_or_else(|_| url_str)
     }
 }
