+++ chain-signatures/node/src/rpc.rs
@@ -414,7 +414,7 @@ impl NearClient {
         self.client
             .call(&self.signer, &self.contract_id, "join")
             .args_json(json!({
-                "url": self.my_addr,
+                "url": crate::util::process_url(self.my_addr.to_string()),
                 "cipher_pk": self.cipher_pk.to_bytes(),
                 "sign_pk": self.sign_pk,
             }))
