+++ chain-signatures/node/src/util.rs
@@ -95,3 +95,20 @@ pub fn is_url_host_ip(url: &Url) -> bool {
         .map(|host| host.parse::<IpAddr>().is_ok())
         .unwrap_or(false)
 }
+
+pub fn process_url(url_str: String) -> String {
+    let url_str = url_str.clone();
+    url::Url::parse(&url_str)
+        .map(|mut url| {
+            if !is_url_host_ip(&url) {
+                if let Err(err) = url.set_port(None) {
+                    tracing::warn!("Error setting participant's url {url} port to None: {err:?}");
+                    return url_str.clone();
+                }
+                url.to_string()
+            } else {
+                url_str.clone()
+            }
+        })
+        .unwrap_or_else(|_| url_str)
+}
